Reworked component data management and event listening

+++ CORE/C_Instances/QuadTree.cpp
@@ -89,10 +89,10 @@ void QuadTree::parentCheck(std::vector<Collidable*> objectsToCheck, QuadTree* pa
 //Check all elements in the vector against all other elements in the vector
 void QuadTree::selfCheck(std::vector<Collidable*> vectorToCheck)
 {
-	for (int i = 0; i < vectorToCheck.size(); i++)
+	for (int i = 0; i < (int) vectorToCheck.size(); i++)
 	{
 		Collidable* obj1 = vectorToCheck[i];
-		for (int j = 0; j < vectorToCheck.size(); j++)
+		for (int j = 0; j < (int) vectorToCheck.size(); j++)
 		{
 			Collidable* obj2 = vectorToCheck[j];
 			
@@ -115,10 +115,10 @@ void QuadTree::selfCheck(std::vector<Collidable*> vectorToCheck)
 
 void QuadTree::checkAgainst(std::vector<Collidable*> v1, std::vector<Collidable*> v2)
 {
-	for (int i = 0; i < v1.size(); i++)
+	for (int i = 0; i < (int) v1.size(); i++)
 	{
 		Collidable* obj1 = v1[i];
-		for (int j = 0; j < v2.size(); j++)
+		for (int j = 0; j < (int) v2.size(); j++)
 		{
 			Collidable* obj2 = v2[j];
 			if (i != j)
@@ -160,7 +160,7 @@ void QuadTree::insert(Collidable* objectToAdd)
 //Remove a Collidable from the current QuadTree
 void QuadTree::remove(Collidable* objectToRemove)
 {
-	for (int i = 0; i < objectsInTree.size(); i++)
+	for (int i = 0; i < (int) objectsInTree.size(); i++)
 	{
 		if (objectsInTree[i] == objectToRemove)
 		{
